+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ## 2017-10-06
 
 - Initial release.
+
+## 2017-10-19
+
+- Updated filter python code documentation.